Novo update nas rotas

+++ src/controllers/controllerUser/controllerDelete.ts
@@ -7,6 +7,18 @@ export class controllerDelete {
   async handle(request: Request, response: Response) {
     const id = request.params.id;
 
+    const idExists = await prisma.user.findUnique({
+      where: {
+        id: id
+      }
+    })
+
+    if(!idExists){
+      return response.status(400).json({
+        msg: `Esse id: ${id} não existe mais no database`
+      })
+    }  
+
     const user = await prisma.user.delete({     
         where: {
           id: id
@@ -20,6 +32,18 @@ export class controllerDelete {
 export class controllerDeleteRelation {
   async handle(request: Request, response: Response) {
     const { id } = request.params;
+
+    const idExists = await prisma.user.findUnique({
+      where: {
+        id: id
+      }
+    })
+
+    if(!idExists){
+      return response.status(400).json({
+        msg: `Esse id: ${id} não existe mais no database`
+      })
+    }  
 
     const user = await prisma.relationsAdress.delete({     
         where: {
